Remove Clock, incapsulate Realm in Database, NoteConformable protocol
Remove Clock class because Swift has already implemented same
functionality with DateFormatter class. Incapsulate DateFormatter inside of update(withText:) method in Note class, so others can update the text of the note and don't bother about time and date updates

Incapsulate all operations with Realm database inside of Database class
and provide public interface for all the operations. If we will need to
change database on some another, for example SQLite or Core Data, we
easilly do this just replacing the database's class in database property and editing all methods that make public interface

Implement NoteConformable protocol and assign it to Note class.
NoteComformable is called to make ShortNote independent of type that he will be initialized from

Add scrollDownIfTextCrossOverKeyboard() method. When a user will be typing a text of note and it will be too large to fit in current view the function allows automatically scroll down and continue typing creating a positive user experience