Provide 'database' property for each class which use the database
Provide 'database' property for each class using the database. In the past, few classes, such as 'AppDelegate' or 'CreateNoteViewController', were using the 'database' property of 'NotesListViewController' in order to perform app related operations which architecturally creates unwanted dependencies on each other. Resolve this by providing a separate 'database' properties in the class declaration and using them inside of the classes' implementations to unbound them.